+++ packages/create-component/build.config.js
@@ -1,0 +1,18 @@
+import { defineBuildConfig } from "unbuild";
+
+export default defineBuildConfig({
+	entries: ["src/index"],
+  outDir: "bin",
+	clean: true,
+	rollup: {
+		inlineDependencies: true,
+		esbuild: {
+			target: "node18",
+			minify: true
+		}
+	},
+	alias: {
+		// we can always use non-transpiled code since we support node 18+
+		prompts: "prompts/lib/index.js"
+	}
+});
