+++ packages/create-component/template/react/gitignore
