--- packages/tester/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "tester",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@kissflow/create-kf-test-comp": "workspace:^"
-  }
-}
