+++ pnpm-workspace.yaml
@@ -1,0 +1,4 @@
+packages:
+  - 'docs/'
+  # all packages in direct subdirs of packages/
+  - 'packages/*'
