--- pnpm-workspaces.yaml
@@ -1,4 +1,0 @@
-packages:
-  # all packages in direct subdirs of packages/
-  - 'docs/'
-  - 'packages/*'
